requirements.txt
antialiased_cnns==0.3
apex==0.9.10dev
cuml==0.6.1.post1
einops==0.8.0
fvcore==0.1.5.post20221221
matplotlib==3.7.1
mmcv==2.0.0
mmcv_full==1.6.1
mmdet==3.3.0
mmsegmentation==0.27.0
numpy==1.24.2
omegaconf==2.3.0
opencv_python==4.5.5.64
opencv_python_headless==4.8.1.78
Pillow==9.5.0
Pillow==5.1.0
Pillow==10.3.0
scipy==1.9.1
seaborn==0.13.2
submitit==1.5.1
timm==0.3.2
torch==1.12.1+cu113
torchmetrics==1.2.0
torchvision==0.13.1+cu113
xformers.egg==info
